requirements.txt
mysql-connector-python
dotenv
flask